feat(cache): invalidate user cache

+++ src/modules/user/service.ts
@@ -1,10 +1,12 @@
-import { UnAuthorizedException } from 'http-exception-transformer/exceptions'
+import { NotFoundException, UnAuthorizedException } from 'http-exception-transformer/exceptions'
 import UserSchema from './model'
 import { CreateUserInterface, UserInterface } from './interface'
 import { Cache } from '../../services/cache/decorator'
+import { invalidateUser } from './cache'
 
 class UserService {
   /** to list all users */
+  @Cache()
   static async findAll(): Promise<Array<UserInterface>> {
     const users = await UserSchema.find().select('-_id -__v -password')
     return users
@@ -24,19 +26,27 @@ class UserService {
     if (user === null) {
       throw new UnAuthorizedException('Invalid Credentials')
     }
+
     return user
   }
 
   /** to create a new user */
   static async create(user: CreateUserInterface): Promise<UserInterface> {
     const newUser = await UserSchema.create(user)
+    invalidateUser(newUser.email, newUser.password)
     delete newUser.password
     return newUser
   }
 
   /** to delete a user by id */
   static async deleteUserByEmail(email: string): Promise<any> {
+    const userData = await this.findOneByEmail(email)
+    if (userData === null) {
+      throw new NotFoundException('User Not Found')
+    }
+
     const deleteResponse = await UserSchema.deleteOne({ email })
+    invalidateUser(userData.email, userData.password)
     return deleteResponse
   }
 }
